feat: add login to delete user

+++ queries-and-resolvers/src/graphql/user/userResolver.ts
@@ -1,31 +1,36 @@
+import { AuthenticationError } from 'apollo-server'
+
 const users = async (_, { input }, { dataSources }) => {
-  const users = await dataSources.userApi.getUsers(input);
-  return users;
-};
+  const users = await dataSources.userApi.getUsers(input)
+  return users
+}
 
 const user = async (_, { id }, { dataSources }) => {
-  const user = await dataSources.userApi.getUser(id);
-  return user;
-};
+  const user = await dataSources.userApi.getUser(id)
+  return user
+}
 
 const createUser = async (_, { data }, { dataSources }) => {
-  return dataSources.userApi.createUser(data);
-};
+  return dataSources.userApi.createUser(data)
+}
 
 const updateUser = async (_, { userId, data }, { dataSources }) => {
-  return dataSources.userApi.updateUser(userId, data);
-};
+  return dataSources.userApi.updateUser(userId, data)
+}
 
-const deleteUser = async (_, { userId }, { dataSources }) => {
-  return dataSources.userApi.deleteUser(userId);
-};
+const deleteUser = async (_, { userId }, { dataSources, loggedUserId }) => {
+  if (!loggedUserId) {
+    throw new AuthenticationError('You have to login')
+  }
+  return dataSources.userApi.deleteUser(userId)
+}
 
 const posts = ({ id }, _, { dataSources }) => {
-  return dataSources.postApi.batchLoadByUserId(id);
-};
+  return dataSources.postApi.batchLoadByUserId(id)
+}
 
 export const userResolvers = {
   Query: { user, users },
   Mutation: { createUser, updateUser, deleteUser },
   User: { posts },
-};
+}
